repository & memento Employee

+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/entity/statutory/worktime_new/employee/KshstShaFlexSetPK.java
@@ -35,7 +35,7 @@ public class KshstShaFlexSetPK implements Serializable {
 	
 	/** The year. */
 	@Column(name = "YEAR")
-	private short year;
+	private int year;
 
 	/* (non-Javadoc)
 	 * @see java.lang.Object#hashCode()
